requirements.txt
boto3
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
python-dotenv
pydantic
email-validator
python-multipart
passlib[bcrypt]
redis